# event-manager/src/main/java/ru/haritonenko/eventmanager/user/api/dto/UserDto.java
package ru.haritonenko.eventmanager.user.api.dto;

import com.fasterxml.jackson.annotation.JsonInclude;
import lombok.Builder;

@Builder
@JsonInclude(JsonInclude.Include.NON_NULL)
public record UserDto(
        Integer id,
        String login
) {
}